Convert updatable to a decorator (and to TS)
